Fix signed/unsigned mismatch in object affect modifiers
Object affect modifiers (e.g., -2 to hitroll) were stored as unsigned
types, causing negative values to become huge positive numbers (~4
billion) when loaded on 64-bit systems. This triggered UBSan runtime
errors when the values overflowed into signed char/short fields.

Changes:
- obj_affected_type.modifier: unsigned long -> long
- obj_affected_type_file.modifier: compat_ulong -> compat_long
- Use COMPAT_TO_LONG/LONG_TO_COMPAT for proper sign extension
- Remove incorrect (unsigned) cast in db.c reset_char()
- Remove now-redundant (long) casts in handler.c

Binary file compatibility is preserved since compat_long and
compat_ulong are both 32-bit types with identical size/alignment.

+++ code/db.c
@@ -1366,7 +1366,7 @@ struct obj_data* read_object(int nr, int type) {
     fscanf(obj_f, " %d ", &tmp);
     obj->affected[i].location = (short)tmp;
     fscanf(obj_f, " %d \n", &tmp);
-    obj->affected[i].modifier = (unsigned long)tmp;
+    obj->affected[i].modifier = (long)tmp;
     fscanf(obj_f, " %s \n", chk);
   }
 
@@ -2357,7 +2357,7 @@ void reset_char(struct char_data* ch) {
   */
 
   for (af = ch->affected; af; af = af->next) {
-    affect_modify(ch, af->location, (unsigned)af->modifier, af->bitvector, 1);
+    affect_modify(ch, af->location, af->modifier, af->bitvector, 1);
   }
 
   if (!HasClass(ch, CLASS_MONK)) {
